fix: billable services side nav issue
