+++ packages/esm-billing-app/src/billable-services/billable-services-home.component.tsx
@@ -3,39 +3,18 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 import { BillableServicesDashboard } from './dashboard/dashboard.component';
 import AddBillableService from './create-edit/add-billable-service.component';
-import { SideNav, SideNavItems, SideNavLink } from '@carbon/react';
 import styles from './billable-services.scss';
 import BillingHeader from '../billing-header/billing-header.component';
-import { Wallet, Money } from '@carbon/react/icons';
-import { UserHasAccess, navigate } from '@openmrs/esm-framework';
 import BillManager from './bill-manager/bill-manager.component';
-
-const basePath = `${window.spaBase}/billable-services`;
+import { BillableServicesSideNav } from './billable-services-sidenav.component';
 
 const BillableServiceHome: React.FC = () => {
   const { t } = useTranslation();
 
-  const handleNavigation = (path: string) => {
-    navigate({ to: `${basePath}/${path}` });
-  };
-
   return (
     <BrowserRouter basename={`${window.spaBase}/billable-services`}>
       <main className={styles.mainSection}>
-        <section>
-          <SideNav>
-            <SideNavItems>
-              <SideNavLink onClick={() => handleNavigation('')} renderIcon={Wallet} isActive>
-                {t('billableServices', 'Billable Services')}
-              </SideNavLink>
-              <UserHasAccess privilege="coreapps.systemAdministration">
-                <SideNavLink onClick={() => handleNavigation('patient-bill-manager')} renderIcon={Money}>
-                  {t('billManager', 'Bill Manager')}
-                </SideNavLink>
-              </UserHasAccess>
-            </SideNavItems>
-          </SideNav>
-        </section>
+        <BillableServicesSideNav />
         <section>
           <BillingHeader title={t('billServicesManagement', 'Bill services management')} />
           <Routes>
